[refactor] interface 분리
